refactor: move validators eb info to own file
